requirements.txt
mysql-connector-python
lxml
jxmlease

